Close out the 50/500 code-quality milestone: bin/lint fully green
The file-decomposition wave landed via the merge queue (which runs pytest, not bin/lint),
so its new modules carried lint debris. This close-out makes bin/lint fully GREEN at the
current thresholds — functions ≤50, files ≤500, args≤4 (relaxed per-file for math/
render-dense), CC≤10, plus coupling/LCOM4/DIT/NOC — reaching the 50/500 milestone.

- Clear wave debris: 194 F401/I001, E402, PLW0108, RUF046, PLR0911; relax PLR0913
  per-file for the app.py-descendant render modules.
- Exempt grid_tables.py (irreducible grid data tables) from the file-length cap, with a
  test pinning the exempt set to exactly the three data modules so no logic module slips in.
- Defer E501 + ruff format together on the dense view modules: wrapping their long
  cell/SVG lines to ≤100 chars busts a length cap, which only the parked 500→250→100
  decomposition can resolve. Format stays enforced on every other module.
- Decompose rendering.py and spreadsheet_geometry.py (the two logic files over 500) into
  cohesive mixins: _ChromeSyncMixin and _ClosedFormMixin.
- Restore app.py's test-facing re-export facade (F401-exempt like __init__.py).
- Document the milestone + the deferred ratchet in CODE_QUALITY.md.

+++ rtt/app/render_html_glyphs.py
@@ -85,7 +85,7 @@ def _chart_ticks(lo: float, hi: float) -> list[float]:
     step = next(m * mag for m in (1, 2, 2.5, 5, 10) if raw <= m * mag)
     start = math.floor(lo / step) * step
     stop = (math.floor(hi / step) + 1) * step
-    count = int(round((stop - start) / step))
+    count = round((stop - start) / step)
     ticks = [round(start + i * step, 6) for i in range(count + 1)]
     if len(ticks) < 2 or ticks[-1] == ticks[0]:
         return [ticks[0], ticks[0] + 1.0]
